Feat: Pagination 구현 완료!!!!!

+++ Task-2/src/main/java/com/management/todoapp/todo/controller/TodoController.java
@@ -1,13 +1,18 @@
 package com.management.todoapp.todo.controller;
 
+import com.management.todoapp.shared.domain.Pageable;
 import com.management.todoapp.todo.dto.request.RequestModifyTodoDto;
 import com.management.todoapp.todo.dto.request.RequestPasswordDto;
 import com.management.todoapp.todo.dto.request.RequestTodoDto;
+import com.management.todoapp.todo.dto.response.ResponseTodoDto;
 import com.management.todoapp.todo.service.TodoService;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequestMapping("/v1/api/todo")
@@ -17,7 +22,13 @@ public class TodoController implements TodoApiDocs{
 
     @Override
     @PostMapping("")
-    public ResponseEntity<?> addTodo(@RequestBody RequestTodoDto requestTodoDto) {
+    public ResponseEntity<?> addTodo(@Valid @RequestBody RequestTodoDto requestTodoDto) {
+        if(requestTodoDto.title().length() > 200){
+            return ResponseEntity.badRequest().body("200자를 넘으면 안됩니다.");
+        }
+        if(requestTodoDto.password().isEmpty()){
+            return ResponseEntity.badRequest().body("비밀번호가 필수입니다.");
+        }
         todoService.createTodo(requestTodoDto);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
@@ -29,15 +40,22 @@ public class TodoController implements TodoApiDocs{
         return ResponseEntity.ok(todoService.getAllTodos());
     }
 
-    /*
     // Pagination 처리
-
     @Override
-    @GetMapping("")
-    public ResponseEntity<?> getTodos() {
-        return null;
+    @GetMapping("/pagination")
+    public ResponseEntity<?> getTodosPagination(
+            @RequestParam int page,
+            @RequestParam int size,
+            @RequestParam(required = false) Integer authorId,
+            @RequestParam(required = false) String updatedAt
+    ) {
+        if(page < 0 || size < 0){
+            return ResponseEntity.ok(List.of());
+        }
+        Pageable pageable = new Pageable(authorId, updatedAt, size, page);
+        List<ResponseTodoDto> todo = todoService.getPagingTodos(pageable);
+        return ResponseEntity.ok(todo);
     }
-    */
 
     @Override
     @GetMapping("/{id}")
@@ -51,6 +69,9 @@ public class TodoController implements TodoApiDocs{
             @PathVariable String id,
             @RequestBody RequestPasswordDto requestPasswordDto
     ) {
+        if(requestPasswordDto.password().isEmpty()){
+            return ResponseEntity.badRequest().body("비밀번호가 필수입니다.");
+        }
         todoService.deleteTodo(id, requestPasswordDto);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
@@ -61,6 +82,9 @@ public class TodoController implements TodoApiDocs{
             @PathVariable String id,
             @RequestBody RequestModifyTodoDto requestModifyTodoDto
     ) {
+        if(requestModifyTodoDto.password().isEmpty()){
+            return ResponseEntity.badRequest().body("비밀번호가 필수입니다.");
+        }
         todoService.updateTodo(id, requestModifyTodoDto);
         return new ResponseEntity<>(HttpStatus.OK);
     }
